requirements.txt
aiogram==3.22.0
aiohttp==3.10.10
aiosqlite==0.20.0
APScheduler==3.10.4
python-dotenv==1.0.1